// src/interfaces/product.interface.ts
import type { User } from "./user.interface";

export interface Product {
    id:          string;
    title:       string;
    price:       number;
    description: string;
    slug:        string;
    stock:       number;
    sizes:       Size[];
    gender:      Gender;
    tags:        string[];
    images:      string[];
    user:        User;
}

export type  Size = "XXS" | "XS" | "S" | "M" | "L" | "XL" | "XXL" | "XXXL";

export type Gender = "kid" | "men" | "women" | "unisex";